Create Movie_AI
Attempt at creating a RAG app that would use IMDB as the database and be able to answer questions that were relevant to movies/actors within the given database. I had two issues in trying to get this to run, the first being that I was consistently having a rate quota limit issue from open AI, I ran into this any time that I tried to use openAI in VS code. The second problem I had was getting the the titles of the rows within the dataset to line up with the ones I had in my code, there were many different names within the five Non-Commercial data sets and getting them all to line up became an issue for me. I then decided to instead create a pdf reader that the user could ask questions to that were relevant to the given pdf that the user would be able to upload.