Create RAG_PDF
This was my attempt at creating a RAG app that would be able to answer questions relevant to a given PDF of a textbook. I built a working one that used hugging face and GPT2 but it was incredibly slow so I tried using openAI for this version and ran into the same Rate Quota Limit issue.