Create Textbook_AI_Streamlit
In this attempt I tried to use langchain and streamlit which is an open source framework which is supposed to help you to build and share machine learning apps. I did my best to try and follow the commands that are given with streamlit but I could not get the command line to function in the proper way to get the model running.